редактирование первого теста согласно требованиям
